Create reusable ZetkinAutoTextArea component
