+++ src/components/ZetkinAutoTextArea.tsx
@@ -1,0 +1,40 @@
+import React from 'react';
+import { lighten, makeStyles, TextareaAutosize } from '@material-ui/core';
+
+const useStyles = makeStyles((theme) => ({
+  textarea: {
+    border: '2px dotted transparent',
+    borderColor: lighten(theme.palette.primary.main, 0.65),
+    borderRadius: 10,
+    fontFamily: theme.typography.fontFamily,
+    lineHeight: '1.5',
+    overflow: 'hidden',
+    padding: 10,
+    resize: 'none',
+    width: '100%',
+  },
+}));
+
+interface ZetkinAutoTextAreaProps {
+  value: string;
+  onChange: (value: string) => void;
+}
+
+const ZetkinAutoTextArea = React.forwardRef<
+  HTMLTextAreaElement,
+  ZetkinAutoTextAreaProps
+>(function ZetkinAutoTextArea({ onChange, value, ...restProps }, ref) {
+  const classes = useStyles();
+
+  return (
+    <TextareaAutosize
+      ref={ref}
+      className={classes.textarea}
+      onChange={(e) => onChange(e.target.value)}
+      value={value}
+      {...restProps}
+    />
+  );
+});
+
+export default ZetkinAutoTextArea;
